+++ changelog.md
@@ -1,5 +1,5 @@
-[date] 
-*For announcements and more info on updates, join the Support Server: https://discord.gg/z76mn4r*
+[date]
+*For announcements and more info on updates, join the [Support Server](https://discord.gg/z76mn4r)*
 Update [date] V1.6
 
 For this update, I wanted ot go all out. I decided to completely remake all the embed messages to make them look much better. 
@@ -7,6 +7,7 @@ For this update, I wanted ot go all out. I decided to completely remake all the 
 --ADDITIONS--
 - `>privacy` will give you the privacy policy.
 - `clean [amount]` Will remove a specified amount of message.
+- `clear_data` Will clear AdminBOT data and AdminBOT will leave the guild.
 - Permissions warning (AdminBOT will now warn if he has missing perms.)
 
 --UPDATES--